requirements.txt
# Core dependencies
google-genai>=1.0.0
python-dotenv>=1.0.0

# Video processing (for perception agent)
opencv-python>=4.8.0
numpy>=1.24.0

# Optional: Audio processing
# pydub>=0.25.1
# simpleaudio>=1.0.4